added contract
